% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtopgene.R
\name{runtopgene}
\alias{runtopgene}
\title{Run topGene API and convert it to enrichResult}
\usage{
runtopgene(
  genelist,
  categories = c("GeneOntologyMolecularFunction", "GeneOntologyBiologicalProcess",
    "GeneOntologyCellularComponent")
)
}
\arguments{
\item{genelist}{vector of genelist}

\item{categories}{GO categories to filter results by. Defaults to Molecular Function, Biological Process and}
}
\description{
Run topGene API and convert it to enrichResult
}
